This is a diff report showing the changes made to several files in a React application. Here's a summary of the changes:
**Components:**

1. `AddonDetailsContent.tsx`: Minor updates to handle external links and display additional information about the addon, such as authors, creation date, update date, and license.
2. `AddonDetailsFooter.tsx`: Added an avatar component to display author images.

**Index files:**

1. `src/components/features/addons/addon-details/index.ts`: Created a new file that reexports other components in this directory.
2. `src/routes/index.tsx`: Updated the import statement for `React` to use the correct syntax (`import { lazy, Suspense } from 'react';`).

**Other changes:**

1. The `AddonDetailsHeader` component now receives additional props for downloads, follows, icon, and other metadata.
2. The `totalDownload` variable is calculated based on the download counts from both CurseForge and Modrinth.

Overall, these changes aim to improve the display of addon information and metadata in the application's UI.
