+++ src/components/features/addons/AddonDetails.tsx
@@ -1,31 +1,19 @@
 import { Card, CardContent} from '@/components/ui/card';
 import { Skeleton } from '@/components/ui/skeleton';
-import {
-  Github,
-  Globe,
-  Bug,
-} from 'lucide-react';
-import {ReactNode, useEffect} from 'react';
 import { useParams } from 'react-router-dom';
 import { useFetchAddon } from '@/api';
-
 import {CurseForgeAddon, ModrinthAddon} from "@/types";
-import AddonDetailsHeader from "@/components/features/addons/addon-details/AddonDetailsHeader.tsx";
-import AddonDetailsContent from "@/components/features/addons/addon-details/AddonDetailsContent.tsx";
-import AddonDetailsGallery from "@/components/features/addons/addon-details/AddonDetailsGallery.tsx";
-import AddonDetailsDescription from "@/components/features/addons/addon-details/AddonDetailsDescription.tsx";
-import AddonDetailFooter from "@/components/features/addons/addon-details/AddonDetailsFooter.tsx";
+import {
+    AddonDetailFooter,
+    AddonDetailsContent,
+    AddonDetailsDescription,
+    AddonDetailsGallery,
+    AddonDetailsHeader
+} from "@/components/features/addons/addon-details";
 
 export default function AddonDetails() {
   const { slug } = useParams();
   const { data: addon, isLoading, error } = useFetchAddon(slug);
-
-  useEffect(() => {
-    if (!slug) {
-      return;
-    }
-  }, [slug]);
-
 
   if (error) {
     return (
@@ -66,36 +54,16 @@ export default function AddonDetails() {
       : addon.curseforge_raw as CurseForgeAddon;
 
   const gallery = modrinthData?.gallery ?? [];
-
-  const createLink = (icon: ReactNode, label: string, url?: string) => {
-    return url ? { icon, label, url } : null;
-  };
-
-
-
-    const { links } = curseforgeData || {};
-  const { sourceUrl, issuesUrl, websiteUrl } = links || {};
-
-  const externalLinks = [
-    createLink(<Github className="h-4 w-4" />, "Source Code", sourceUrl || ''),
-    createLink(<Bug className="h-4 w-4" />, "Issue Tracker", issuesUrl || ''),
-    createLink(<Globe className="h-4 w-4" />, "Website", websiteUrl),
-  ].filter((link): link is { icon: ReactNode; label: string; url: string } => link !== null);
-
-
-
-
+  const totalDownload = modrinthData?.downloads | 0 + curseforgeData?.downloadCount | 0
   return (
     <div className='container mx-auto space-y-8 px-4 py-8'>
-
       <Card>
-
         {/* Header Card */}
 
       <AddonDetailsHeader
           title={addon.name}
           description={addon.description}
-          downloads={modrinthData?.downloads || 0}
+          downloads={totalDownload}
           follows={modrinthData?.follows || 0}
           icon={addon.icon}
       />
@@ -106,10 +74,9 @@ export default function AddonDetails() {
           versions={addon.minecraft_versions || []}
           loaders={addon.loaders || []}
           categories={addon.categories || []}
-          links={externalLinks || []}
           slug={addon.slug}
-          modrinth_raw={addon.modrinth_raw as ModrinthAddon}
-          curseforge_raw={addon.curseforge_raw as CurseForgeAddon}
+          modrinth_raw={modrinthData}
+          curseforge_raw={curseforgeData}
       />
 
       </Card>
@@ -129,7 +96,9 @@ export default function AddonDetails() {
       />
 
       {/* Additional Information */}
+
       <AddonDetailFooter authors={curseforgeData.authors}  createdAt={addon.created_at} updatedAt={addon.updated_at} licence={modrinthData?.license} addon_name={addon.name}></AddonDetailFooter>
+
     </div>
   );
 }
