+++ src/components/features/addons/addon-details/index.ts
@@ -1,0 +1,6 @@
+// src/components/features/addons/addon-details/index.ts
+export { AddonDetailsHeader } from './AddonDetailsHeader';
+export { AddonDetailsGallery } from './AddonDetailsGallery';
+export { AddonDetailsContent } from './AddonDetailsContent';
+export { AddonDetailFooter } from './AddonDetailsFooter';
+export { AddonDetailsDescription } from './AddonDetailsDescription';
